Add version command
Adds `<prefix>version` which shows the compiled version of Seabird. Also
renames "uptime" to "introspection". Happy to move `version` to a
different plugin and revert the rename if you like!
