+++ src/plugins/introspection.rs
@@ -1,0 +1,57 @@
+use std::fmt::Write;
+
+use time::Instant;
+
+use crate::prelude::*;
+
+pub struct IntrospectionPlugin {
+    started: Instant,
+}
+
+impl IntrospectionPlugin {
+    pub fn new() -> Self {
+        IntrospectionPlugin {
+            started: Instant::now(),
+        }
+    }
+}
+
+const SEABIRD_VERSION: &'static str = env!("CARGO_PKG_VERSION");
+
+#[async_trait]
+impl Plugin for IntrospectionPlugin {
+    fn new_from_env() -> Result<Self> {
+        Ok(IntrospectionPlugin::new())
+    }
+
+    async fn run(self, mut stream: Receiver<Arc<Context>>) -> Result<()> {
+        while let Some(ctx) = stream.next().await {
+            match ctx.as_event() {
+                Event::Command("uptime", _) => {
+                    let elapsed = self.started.elapsed();
+
+                    let days = elapsed.whole_days();
+                    let hours = elapsed.whole_hours() - (days * 24);
+                    let minutes = elapsed.whole_minutes() - (days * 24 * 60) - (hours * 60);
+
+                    let mut ret = String::new();
+
+                    if days > 0 {
+                        write!(ret, "{} days ", days).unwrap();
+                    }
+
+                    write!(ret, "{:02}:{:02}", hours, minutes).unwrap();
+
+                    ctx.mention_reply(&ret[..]).await?;
+                }
+                Event::Command("version", _) => {
+                    ctx.mention_reply(&format!("Seabird version {}", SEABIRD_VERSION))
+                        .await?;
+                }
+                _ => {}
+            }
+        }
+
+        Err(format_err!("introspection plugin exited early"))
+    }
+}
